Rename main parsing fn, break out fn into their separate files
